pyproject.toml
[tool.poetry]
name = "capstone-1-project-zoomcamp"
version = "0.1.0"
description = ""
authors = ["SchneiderSix <104461916+SchneiderSix@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
scikit-learn = "^1.6.0"
pandas = "^2.2.3"
numpy = "2.0.0"
seaborn = "^0.13.2"
matplotlib = "^3.10.0"
flasgger = "^0.9.7.1"
tensorflow = "^2.18.0"
kagglehub = "^0.3.6"
pillow = "^11.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
